requirements.txt
pandas
duckduckgo-search
openbb
pandas-ta
feedparser
python-dotenv
requests
streamlit
Pillow
argparse
langgraph
langchain
langchain-core
langchain-groq
